Finish project
I couldn't figure out a way to remove eventListenerEvent for the black
mode. looks like removeEventListener requires the function to be just
a name and not the whole function written out and I couldn't figure
out how to have it not require arguments.

Goal was to be able to click back and forth between color mode and
gray mode.